Fix empty body php requests

+++ phpcgi.go
@@ -19,30 +19,32 @@ func Cgi(w http.ResponseWriter, r *http.Request, phpBin string, scriptFileName s
 		r.TransferEncoding = r.TransferEncoding[1:]
 	}
 
-	// Save body into temp file (don't want large request bodies eating memory)
-	tmpfile, err := os.CreateTemp("", "fp-cgi-")
-	if err != nil {
-		http.Error(w, "Error creating temp file for request body", http.StatusInternalServerError)
-		return
-	}
-	defer os.Remove(tmpfile.Name())
+	if r.Body != nil {
+		// Save body into temp file (don't want large request bodies eating memory)
+		tmpfile, err := os.CreateTemp("", "fp-cgi-")
+		if err != nil {
+			http.Error(w, "Error creating temp file for request body", http.StatusInternalServerError)
+			return
+		}
+		defer os.Remove(tmpfile.Name())
 
-	// Write the body to the temp file
-	size, err := io.Copy(tmpfile, r.Body)
-	if err != nil {
-		http.Error(w, "Error writing request body to temp file", http.StatusInternalServerError)
-		return
-	}
+		// Write the body to the temp file
+		size, err := io.Copy(tmpfile, r.Body)
+		if err != nil {
+			http.Error(w, "Error writing request body to temp file", http.StatusInternalServerError)
+			return
+		}
 
-	// Move seek to start of file
-	_, err = tmpfile.Seek(0, 0)
-	if err != nil {
-		http.Error(w, "Error seeking to start of temp file", http.StatusInternalServerError)
-		return
-	}
+		// Move seek to start of file
+		_, err = tmpfile.Seek(0, 0)
+		if err != nil {
+			http.Error(w, "Error seeking to start of temp file", http.StatusInternalServerError)
+			return
+		}
 
-	r.Body = tmpfile
-	r.ContentLength = size
+		r.Body = tmpfile
+		r.ContentLength = size
+	}
 
 	handler.ServeHTTP(w, r)
 }
